migrate the files and add docs in google platform
